chore: change all byte types to int and move inventory storage to character

+++ Assets/1_FeedTheRealm/Core/Client/EventChannels/Status/LastItemChangedEvent.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 namespace FTR.Core.Client.EventChannels.Status
 {
     [CreateAssetMenu(menuName = "Events/Client/Status/LastItemChanged")]
-    public class LastItemChangedEvent : EventChannelSO<(string, byte)> { }
+    public class LastItemChangedEvent : EventChannelSO<(string, int)> { }
 }
